+++ internal/mode/validate/column_spec.go
@@ -36,14 +36,10 @@ type schemaRenderer struct {
 	Output *strings.Builder
 }
 
-func (r *schemaRenderer) VisitSet(s variable.Set) {
-	for i, m := range s.Literals() {
-		if i > 0 {
-			r.Output.WriteString(" | ")
-		}
-
-		r.Output.WriteString(m.Quote())
-	}
+func (r *schemaRenderer) VisitBinary(s variable.Binary) {
+	r.Output.WriteByte('<')
+	r.Output.WriteString(s.EncodingDescription())
+	r.Output.WriteByte('>')
 }
 
 func (r *schemaRenderer) VisitNumeric(s variable.Numeric) {
@@ -75,6 +71,16 @@ func (r *schemaRenderer) VisitNumeric(s variable.Numeric) {
 			"<%s>",
 			s.Type().Kind(),
 		)
+	}
+}
+
+func (r *schemaRenderer) VisitSet(s variable.Set) {
+	for i, m := range s.Literals() {
+		if i > 0 {
+			r.Output.WriteString(" | ")
+		}
+
+		r.Output.WriteString(m.Quote())
 	}
 }